backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
mavsdk>=2.0.0
aiosqlite>=0.19.0
pyyaml>=6.0
pyjwt>=2.8.0
pyserial>=3.5
pyserial-asyncio>=0.6
av>=12.0.0
aiortc>=1.6.0
